+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add possibility to define a softlink target with an overlay template
 - Support defining a symlink with an overlay template. #1303
 - New "localtime" overlay to define the system time zone. #1303
+- Adds `wwctl container <exec|shell> --build=false` to prevent automatically (re)building the container. #1490, #1489
 
 ### Changed